linebot/requirements.txt
black==23.1.0
flake8==6.0.0
isort==5.12.0
mypy==1.1.1
mypy-extensions==1.0.0
flask==3.0.3
openai==0.27.2
python-dotenv==1.0.0
line-bot-sdk==2.4.2
gunicorn==20.1.0